+++ utils/IndexBuffer.cpp
@@ -1,6 +1,9 @@
 #include "IndexBuffer.h"
 #include "Renderer.h"
 
+namespace gl_utils
+{
+
 IndexBuffer::IndexBuffer(const unsigned int* data, unsigned int count) : m_Count(count)
 {
     glGenBuffers(1, &m_RendererID);
@@ -23,3 +26,4 @@ void IndexBuffer::unbind() const
     glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, 0);
 }
 
+}
