+++ utils/VertexBuffer.h
@@ -1,5 +1,8 @@
 #pragma once
 
+namespace gl_utils
+{
+
 class VertexBuffer
 {
     unsigned int m_RendererID;
@@ -11,3 +14,4 @@ public:
     void bind() const;
     void unbind() const;
 };
+}
